# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Документація для API управління користувачами.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Основний сервер API"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Отримати список користувачів",
                "description": "Повертає список користувачів з пагінацією.",
                "operationId": "getUsers",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер сторінки для пагінації",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список користувачів",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "total_pages": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "total_users": {
                                            "type": "integer",
                                            "example": 55
                                        },
                                        "count": {
                                            "type": "integer",
                                            "example": 6
                                        },
                                        "links": {
                                            "properties": {
                                                "next_url": {
                                                    "type": "string",
                                                    "example": "http://rest-api.test/api/v1/users?page=2"
                                                },
                                                "prev_url": {
                                                    "type": "string",
                                                    "example": "null"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "users": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 75
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "User Name"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "post@gmail.com"
                                                    },
                                                    "phone": {
                                                        "type": "string",
                                                        "example": "+380734888999"
                                                    },
                                                    "position": {
                                                        "type": "string",
                                                        "example": "Content manager",
                                                        "nullable": true
                                                    },
                                                    "position_id": {
                                                        "type": "string",
                                                        "example": "2"
                                                    },
                                                    "registration_timestamp": {
                                                        "type": "integer",
                                                        "example": 1732819981
                                                    },
                                                    "photo": {
                                                        "type": "string",
                                                        "example": "http://rest-api.test/storage/avatars/1732819981_6748bc0d1435e.jpg"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string",
                                            "example": "http://rest-api.test/api/v1/users?page=1"
                                        },
                                        "from": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "last_page_url": {
                                            "type": "string",
                                            "example": "http://rest-api.test/api/v1/users?page=10"
                                        },
                                        "next_page_url": {
                                            "type": "string",
                                            "example": "http://rest-api.test/api/v1/users?page=2"
                                        },
                                        "path": {
                                            "type": "string",
                                            "example": "http://rest-api.test/api/v1/users"
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 6
                                        },
                                        "prev_page_url": {
                                            "type": "string",
                                            "example": "null"
                                        },
                                        "to": {
                                            "type": "integer",
                                            "example": 6
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 55
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Додати нового користувача",
                "description": "Створює нового користувача з необов'язковим зображенням.",
                "operationId": "createUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phone",
                                    "position_id",
                                    "photo"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Ім'я користувача, повинно бути від 2 до 60 символів",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "description": "Електронна пошта користувача, має бути валідною згідно з RFC2822",
                                        "type": "string",
                                        "example": "john.doe@example.com"
                                    },
                                    "phone": {
                                        "description": "Номер телефону користувача, має починатись з коду України +380",
                                        "type": "string",
                                        "example": "+380123456789"
                                    },
                                    "position_id": {
                                        "description": "Ідентифікатор посади користувача. Список можна отримати через GET api/v1/positions",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "photo": {
                                        "description": "Фото користувача у форматі JPEG/JPG, мінімальний розмір 70x70px, розмір не більше 5MB",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Користувач успішно створений",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 23
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "New user successfully registered"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Токен закінчився",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The token expired."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Помилка валідації",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "fails": {
                                            "type": "object",
                                            "additionalProperties": [
                                                {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The name must be at least 2 characters."
                                                    },
                                                    "property": "name"
                                                },
                                                {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The email must be a valid email address."
                                                    },
                                                    "property": "email"
                                                },
                                                {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The phone field is required."
                                                    },
                                                    "property": "phone"
                                                },
                                                {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The position id must be an integer."
                                                    },
                                                    "property": "position_id"
                                                },
                                                {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The photo may not be greater than 5 Mbytes."
                                                    },
                                                    "property": "photo"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}